+++ packages/database/src/index.js
@@ -1,1 +1,6 @@
 export { createSupabaseClient, pingDatabase } from "./client.js";
+export {
+  saveMarketSnapshot,
+  saveTechnicalIndicators,
+  createMarketDataRepository,
+} from "./marketDataRepository.js";